modificati alcuni metodi libreriajpa

+++ main/java/org/elis/jpa/LibreriaDaoJpa.java
@@ -1,12 +1,15 @@
 package org.elis.jpa;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 import org.elis.dao.LibreriaDao;
 import org.elis.model.Libreria;
+import org.elis.model.Utente;
 
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.EntityTransaction;
+import jakarta.persistence.NoResultException;
 import jakarta.persistence.Query;
 
 public class LibreriaDaoJpa implements LibreriaDao {
@@ -19,7 +22,7 @@ public class LibreriaDaoJpa implements LibreriaDao {
 	public static LibreriaDaoJpa getInstance() {
 		if(instance == null) {
 			instance = new LibreriaDaoJpa();
-		}
+		} 
 		return instance;
 	}
 
@@ -45,7 +48,11 @@ public class LibreriaDaoJpa implements LibreriaDao {
 		EntityManager em = DaoFactoryJpa.getEntityManager();
 		Query q = em.createQuery("select a from Libreria a where a.nome=:nome");
 		q.setParameter("nome", nome);
-		return (Libreria)q.getSingleResult();
+		try {
+			return (Libreria)q.getSingleResult();
+		}catch (NoResultException e) {
+	        return null;
+	    }
 	}
 
 	@Override
@@ -53,30 +60,62 @@ public class LibreriaDaoJpa implements LibreriaDao {
 		EntityManager em = DaoFactoryJpa.getEntityManager();
 		Query q = em.createQuery("select a from Libreria a where a.libreriaUtente.id=:id_utente");
 		q.setParameter("id_utente", id_utente);
-		return q.getResultList();
+		try {
+			return q.getResultList();
+		}catch (NoResultException e) {
+	        e.printStackTrace();
+	        return null;  
+	    }
 	}
 
 	@Override
 	public Libreria updateNome(long id, String nome) {
 		EntityManager em = DaoFactoryJpa.getEntityManager();
-		EntityTransaction t = em.getTransaction();
-		t.begin();
-		Libreria l = em.find(Libreria.class, id);
-		l.setNome(nome);
-		t.commit();
-		return l;
+		Libreria l = null;
+		LocalDateTime now = LocalDateTime.now();
+		try {
+			l = em.find(Libreria.class, id);
+	    	 if (l == null) {
+	             System.out.println("Libreria non trovata per l'ID: " + id);
+	             return null;
+	         }
+	    	 
+	    	 em.getTransaction().begin();
+	    	 l.setNome(nome);
+	    	 l.setData_ultima_modifica(now);
+	    	 em.getTransaction().commit();
+		}catch (Exception e) {
+	    	if (em.getTransaction().isActive()) {
+	            em.getTransaction().rollback();
+	        }
+	        e.printStackTrace();
+	    } finally {
+	    	 em.close();
+	    }
+	    return l;
 	}
 	
 	@Override
 	public Libreria deleteById(long id) {
 		EntityManager em = DaoFactoryJpa.getEntityManager();
-		EntityTransaction t = em.getTransaction();
-		Libreria l = em.find(Libreria.class, id);
-		t.begin();
-		em.remove(l);
-		t.commit();
-		return l;
-	}
-	
-	
+		Libreria l = null;
+		try {
+			l = em.find(Libreria.class, id);
+			if (l != null) {
+	            em.getTransaction().begin();
+	            em.remove(l);
+	            em.getTransaction().commit();
+	        } else {
+	            System.out.println("Libreria non trovata con ID: " + id);
+	        }
+
+	    } catch (NoResultException e) {
+	        e.printStackTrace();
+	    } catch (Exception e) {
+	    
+	        e.printStackTrace();
+	    } 
+
+	    return l;
+		}
 }
